Integrated java9 code, removing multi-jar
Moved main/java9 code into main/java and removed multi-jar build stuff
